+++ CHANGES.md
@@ -1,5 +1,7 @@
 # 0.3.0
 - Showing LFRP ads
+- Added some thin subtle dividers to the layout to separate things a little better.
+- Added the ability to change channel message display mode (chat, lfrp, both)
 
 # 0.2.1
 - Fixed a nasty bug where the login menu didn't show up.